+++ wgpu-core/src/device/resource.rs
@@ -19,8 +19,8 @@ use smallvec::SmallVec;
 use wgpu_sync::atomic::{AtomicBool, Ordering};
 use wgpu_sync::OnceCell;
 use wgt::{
-    math::align_to, ColorWrites, DeviceLostReason, TextureFormat, TextureSampleType,
-    TextureViewDimension,
+    error::WebGpuError, math::align_to, ColorWrites, DeviceLostReason, TextureFormat,
+    TextureSampleType, TextureViewDimension,
 };
 
 #[cfg(feature = "trace")]
@@ -46,7 +46,7 @@ use crate::{
     },
     instance::{Adapter, RequestDeviceError},
     lock::{rank, Mutex, RwLock},
-    pipeline::{self, ColorStateError},
+    pipeline::{self, shader_module_error_into_compilation_info, ColorStateError},
     pool::ResourcePool,
     resource::{
         self, Buffer, BufferState, ExternalTexture, ExternalTextureState, Labeled, ParentDevice,
@@ -2410,10 +2410,7 @@ impl Device {
         self: &Arc<Self>,
         desc: &pipeline::ShaderModuleDescriptor<'a>,
         source: pipeline::ShaderModuleSource<'a>,
-    ) -> (
-        Arc<pipeline::ShaderModule>,
-        Option<pipeline::CreateShaderModuleError>,
-    ) {
+    ) -> Arc<pipeline::ShaderModule> {
         profiling::scope!("Device::create_shader_module");
         #[cfg(feature = "trace")]
         let data = self.trace.lock().as_mut().map(|trace| {
@@ -2443,14 +2440,17 @@ impl Device {
                 }
             }
         });
-        let (shader, error) = match self.create_shader_module_inner(desc, source) {
-            Ok(shader) => (shader, None),
-            Err(e) => {
-                let shader =
-                    pipeline::ShaderModule::invalid(Arc::clone(self), desc.label.to_string());
-                (shader, Some(e))
-            }
-        };
+        let shader = self
+            .create_shader_module_inner(desc, source)
+            .unwrap_or_else(|e| {
+                let shader = pipeline::ShaderModule::invalid(
+                    Arc::clone(self),
+                    desc.label.to_string(),
+                    shader_module_error_into_compilation_info(&e),
+                );
+                self.handle_error(e, desc.label.as_deref(), "Device::create_shader_module");
+                shader
+            });
         api_log!("Device::create_shader_module -> {:?}", Arc::as_ptr(&shader));
 
         #[cfg(feature = "trace")]
@@ -2468,7 +2468,7 @@ impl Device {
                     data,
                 });
         };
-        (shader, error)
+        shader
     }
 
     pub(crate) fn create_shader_module_inner<'a>(
@@ -2601,6 +2601,7 @@ impl Device {
             }),
             device: self.clone(),
             label: desc.label.to_string(),
+            compilation_info: wgt::CompilationInfo::default(),
         };
 
         let module = Arc::new(module);
@@ -2615,20 +2616,24 @@ impl Device {
     pub unsafe fn create_shader_module_passthrough<'a>(
         self: &Arc<Self>,
         desc: &pipeline::ShaderModuleDescriptorPassthrough<'a>,
-    ) -> (
-        Arc<pipeline::ShaderModule>,
-        Option<pipeline::CreateShaderModuleError>,
-    ) {
+    ) -> Arc<pipeline::ShaderModule> {
         profiling::scope!("Device::create_shader_module_passthrough");
 
-        let (shader, error) = match unsafe { self.create_shader_module_passthrough_inner(desc) } {
-            Ok(shader) => (shader, None),
-            Err(e) => {
-                let shader =
-                    pipeline::ShaderModule::invalid(Arc::clone(self), desc.label.to_string());
-                (shader, Some(e))
-            }
-        };
+        let shader =
+            unsafe { self.create_shader_module_passthrough_inner(desc) }.unwrap_or_else(|e| {
+                let shader = pipeline::ShaderModule::invalid(
+                    Arc::clone(self),
+                    desc.label.to_string(),
+                    shader_module_error_into_compilation_info(&e),
+                );
+                self.handle_error(
+                    e,
+                    desc.label.as_deref(),
+                    "Device::create_shader_module_passthrough",
+                );
+                shader
+            });
+
         #[cfg(feature = "trace")]
         if let Some(ref mut trace) = *self.trace.lock() {
             use crate::device::trace::{DataKind, IntoTrace as _};
@@ -2661,7 +2666,7 @@ impl Device {
             "Device::create_shader_module_spirv -> {:?}",
             Arc::as_ptr(&shader)
         );
-        (shader, error)
+        shader
     }
 
     pub(crate) unsafe fn create_shader_module_passthrough_inner<'a>(
@@ -2762,6 +2767,7 @@ impl Device {
             }),
             device: self.clone(),
             label: descriptor.label.to_string(),
+            compilation_info: wgt::CompilationInfo::default(),
         };
 
         Ok(Arc::new(module))
@@ -4320,7 +4326,7 @@ impl Device {
     ) -> Arc<pipeline::ComputePipeline> {
         profiling::scope!("Device::create_compute_pipeline");
         let compute_pipeline = self
-            .create_compute_pipeline_or_error(desc.clone())
+            .create_compute_pipeline_or_error_inner(desc.clone())
             .unwrap_or_else(|err| {
                 if let pipeline::CreateComputePipelineError::Internal(ref error) = err {
                     log::error!(
@@ -4355,9 +4361,24 @@ impl Device {
     }
 
     /// Creates a compute pipeline without raising any error to device.
+    /// Device lost errors will be mapped to invalid compute pipeline
+    /// as required by specification.
     ///
     /// Corresponds to [GPUDevice.createComputePipelineAsync](https://www.w3.org/TR/webgpu/#dom-gpudevice-createcomputepipelineasync)
     pub fn create_compute_pipeline_or_error(
+        self: &Arc<Self>,
+        desc: pipeline::ComputePipelineDescriptor,
+    ) -> Result<Arc<pipeline::ComputePipeline>, pipeline::CreateComputePipelineError> {
+        let label = desc.label.to_string();
+        match self.create_compute_pipeline_or_error_inner(desc) {
+            Err(err) if err.webgpu_error_type() == wgt::error::ErrorType::DeviceLost => {
+                Ok(pipeline::ComputePipeline::invalid(self.clone(), label))
+            }
+            result => result,
+        }
+    }
+
+    fn create_compute_pipeline_or_error_inner(
         self: &Arc<Self>,
         desc: pipeline::ComputePipelineDescriptor,
     ) -> Result<Arc<pipeline::ComputePipeline>, pipeline::CreateComputePipelineError> {
@@ -4517,7 +4538,7 @@ impl Device {
         profiling::scope!("Device::create_render_pipeline");
 
         let render_pipeline = self
-            .create_render_pipeline_or_error(desc.clone())
+            .create_render_pipeline_or_error_inner(desc.clone())
             .unwrap_or_else(|err| {
                 if let pipeline::CreateRenderPipelineError::Internal { stage, ref error } = err {
                     log::error!("Shader translation error for stage {stage:?}: {error}");
@@ -4542,9 +4563,24 @@ impl Device {
     }
 
     /// Creates a render pipeline without raising any error to device.
+    /// Device lost errors will be mapped to invalid render pipeline
+    /// as required by specification.
     ///
     /// Corresponds to [GPUDevice.createRenderPipelineAsync](https://www.w3.org/TR/webgpu/#dom-gpudevice-createrenderpipelineasync)
     pub fn create_render_pipeline_or_error(
+        self: &Arc<Self>,
+        desc: pipeline::ResolvedGeneralRenderPipelineDescriptor,
+    ) -> Result<Arc<pipeline::RenderPipeline>, pipeline::CreateRenderPipelineError> {
+        let label = desc.label.to_string();
+        match self.create_render_pipeline_or_error_inner(desc) {
+            Err(e) if e.webgpu_error_type() == wgt::error::ErrorType::DeviceLost => Ok(
+                pipeline::RenderPipeline::invalid(self.clone(), label.to_string()),
+            ),
+            result => result,
+        }
+    }
+
+    fn create_render_pipeline_or_error_inner(
         self: &Arc<Self>,
         desc: pipeline::ResolvedGeneralRenderPipelineDescriptor,
     ) -> Result<Arc<pipeline::RenderPipeline>, pipeline::CreateRenderPipelineError> {